ai-server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
chromadb>=0.4.0
chroma-hnswlib==0.7.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
docx2txt>=0.8
durationpy==0.10
faiss-cpu>=1.7.4
fastapi==0.104.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.1
google-api-python-client==2.177.0
googleapis-common-protos==1.70.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.2
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==33.1.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core==0.3.72
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0
langgraph>=0.6.2
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langsmith==0.4.10
lxml==6.0.0
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.2
oauthlib==3.3.1
onnxruntime==1.22.1
openai>=1.10.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
orjson==3.11.1
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
posthog==6.3.3
propcache==0.3.2
protobuf==6.31.1
proto-plus==1.26.1
pyasn1_modules==0.4.2
pyasn1==0.6.1
pydantic_core==2.33.2
pydantic==2.11.7
pydantic-settings==2.10.1
Pygments==2.19.2
pyparsing==3.2.3
pypdf>=3.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx>=0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.31.0 
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.36.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.27.0
sympy==1.14.0
tenacity==9.1.2
tiktoken>=0.5.0
tokenizers==0.21.4
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.1
typing-extensions>=4.8.0
typing-inspect==0.9.0
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn[standard]==0.24.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0